Added a link to the reservation in various emails
Changed .nil? on notes to .blank?
added markdown to the notes that were missing it
(still need to test)

Signed-off-by: Alan Liu <AlanL.Liu@yahoo.com>

Added an absolute link to the reservations in checkin receipt, checkout receipt, checked_in_fine email, and reservation confirmation.

Added tests to those four emails to see if they contain a link with reservations in it.

fixed travis offenses

removed extra spaces, reverted a few changes to requirements notes,

improved the test by adding in the full url. added a custom test for one case because @res is an array

travis

travis try #2